refactored Player class, separating out the ability system functionality into a new class called PlayerAbilityController.  Updated other scripts to work with this. Created a CharacterClass script from which all character classes can derive. Made Knight derive from it.
