+++ ARPG-CSE5912-LTS/Assets/Scripts/Controllers/Player/PlayerAbilityController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6d0e28199727b5147b4057d07db30eec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
