+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/AbilityModel/Ability Cast Type/BaseCastType.cs
@@ -24,7 +24,7 @@ public abstract class BaseCastType : MonoBehaviour
 
     private void OnEnable()
     {
-        Player.AbilityIsReadyToBeCastEvent += OnAbilityIsReadyToBeCast;
+        PlayerAbilityController.AbilityIsReadyToBeCastEvent += OnAbilityIsReadyToBeCast;
         MovementHandler.PlayerBeganMovingEvent += OnPlayerBeganMoving;
     }
