+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/Characters/Classes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 73a4e9b6a3b68b74d82145d8770a120a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
