+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/Characters/Classes/CharacterClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b98d0c81ec95d34e891867df99cb543
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
