+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/Characters/Classes/EnemyKnight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1600f51fb1e1d1e4a9176f2349d2d912
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
