--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/Characters/Knight.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 83e53c75216c4f040baa241b889bbc11
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
